# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Habit Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grocery-items:
    get:
      operationId: listGroceryItems
      tags: [grocery]
      summary: List grocery items for the signed-in user
      responses:
        "200":
          description: Items in display order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createGroceryItem
      tags: [grocery]
      summary: Add a grocery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroceryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /grocery-items/clear-checked:
    post:
      operationId: clearCheckedGroceryItems
      tags: [grocery]
      summary: Delete all checked grocery items
      responses:
        "204":
          description: Cleared
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /grocery-items/{id}:
    patch:
      operationId: updateGroceryItem
      tags: [grocery]
      summary: Update a grocery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroceryItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGroceryItem
      tags: [grocery]
      summary: Delete a grocery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GroceryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 80
        checked:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - checked
        - sortOrder
        - createdAt
    CreateGroceryItemBody:
      type: object
      properties:
        name:
          type: string
          maxLength: 80
        checked:
          type: boolean
          description: Optional; defaults to false. Used when migrating local data.
      required:
        - name
    UpdateGroceryItemBody:
      type: object
      properties:
        name:
          type: string
          maxLength: 80
        checked:
          type: boolean
        sortOrder:
          type: integer